backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
python-dotenv==1.0.1
anthropic==0.40.0
google-generativeai==0.8.3
openai==1.54.5
Pillow==11.0.0
pytesseract==0.3.13
python-docx==1.1.2
pypdf==5.1.0
pyinstaller==6.11.1
